refactor: change package naming from 'helloworld' to 'helloworlddb' and update related configurations
