+++ src/main/java/src/main/java/ch/adeutschmanndev/helloworlddb/HelloworldApplication.java
@@ -1,4 +1,4 @@
-package ch.adeutschmanndev.helloworld;
+package ch.adeutschmanndev.helloworlddb;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
