+++ src/main/java/src/main/java/ch/adeutschmanndev/helloworlddb/model/dto/ErrorResponseDto.java
@@ -1,4 +1,4 @@
-package ch.adeutschmanndev.helloworld.model.dto;
+package ch.adeutschmanndev.helloworlddb.model.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
