+++ src/main/java/src/main/java/ch/adeutschmanndev/helloworlddb/model/dto/UpdateGreetingRequestDto.java
@@ -1,4 +1,4 @@
-package ch.adeutschmanndev.helloworld.model.dto;
+package ch.adeutschmanndev.helloworlddb.model.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
@@ -8,19 +8,19 @@ import lombok.Data;
 
 @Data
 @Builder
-@Schema(description = "Request object for creating a new greeting")
-public class CreateGreetingRequestDto {
+@Schema(description = "Request object for updating an existing greeting")
+public class UpdateGreetingRequestDto {
 
     @NotBlank(message = "Message cannot be blank")
     @Size(max = 500, message = "Message cannot exceed 500 characters")
-    @Schema(description = "The greeting message", example = "Hello World!", required = true)
+    @Schema(description = "The updated greeting message", example = "Hello Updated World!", required = true)
     private String message;
 
     @Size(max = 100, message = "Sender cannot exceed 100 characters")
-    @Schema(description = "The sender of the greeting", example = "alice")
+    @Schema(description = "The updated sender of the greeting", example = "alice")
     private String sender;
 
     @Size(max = 100, message = "Recipient cannot exceed 100 characters")
-    @Schema(description = "The recipient of the greeting", example = "bob")
+    @Schema(description = "The updated recipient of the greeting", example = "bob")
     private String recipient;
 }
