+++ src/main/java/src/main/java/ch/adeutschmanndev/helloworlddb/model/entity/Greetings.java
@@ -1,4 +1,4 @@
-package ch.adeutschmanndev.helloworld.model.entity;
+package ch.adeutschmanndev.helloworlddb.model.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
